Fix an issue where Rush would attempt to open a project's log file for writing twice.

+++ libraries/rush-lib/src/logic/operations/ProjectLogWritable.ts
@@ -60,39 +60,39 @@ export class ProjectLogWritable extends TerminalWritable {
     }
 
     const projectFolder: string = this._project.projectFolder;
-    const {
-      logPath: legacyLogPath,
-      errorLogPath: legacyErrorLogPath,
-      relativeLogPath: legacyRelativeLogPath,
-      relativeErrorLogPath: legacyRelativeErrorLogPath
-    } = getLogFilePaths(projectFolder, 'build');
-    // If the phased commands experiment is enabled, put logs under `rush-logs`
-    if (project.rushConfiguration.experimentsConfiguration.configuration.phasedCommands) {
-      // Delete the legacy logs
-      FileSystem.deleteFile(legacyLogPath);
-      FileSystem.deleteFile(legacyErrorLogPath);
+    // Delete the legacy logs
+    const { logPath: legacyLogPath, errorLogPath: legacyErrorLogPath } = getLogFilePaths(
+      projectFolder,
+      'build'
+    );
+    FileSystem.deleteFile(legacyLogPath);
+    FileSystem.deleteFile(legacyErrorLogPath);
 
+    // If the phased commands experiment is enabled, put logs under `rush-logs`
+    let logFolder: string | undefined;
+    if (project.rushConfiguration.experimentsConfiguration.configuration.phasedCommands) {
       const logPathPrefix: string = `${projectFolder}/${RushConstants.rushLogsFolderName}`;
       FileSystem.ensureFolder(logPathPrefix);
-
-      const { logPath, errorLogPath, relativeLogPath, relativeErrorLogPath } = getLogFilePaths(
-        projectFolder,
-        logFilenameIdentifier,
-        RushConstants.rushLogsFolderName
-      );
-      this.logPath = logPath;
-      this.errorLogPath = errorLogPath;
-      this.relativeLogPath = relativeLogPath;
-      this.relativeErrorLogPath = relativeErrorLogPath;
-    } else {
-      this.logPath = legacyLogPath;
-      this.errorLogPath = legacyErrorLogPath;
-      this.relativeLogPath = legacyRelativeLogPath;
-      this.relativeErrorLogPath = legacyRelativeErrorLogPath;
+      logFolder = RushConstants.rushLogsFolderName;
     }
 
-    FileSystem.deleteFile(this.logPath);
-    FileSystem.deleteFile(this.errorLogPath);
+    const { logPath, errorLogPath, relativeLogPath, relativeErrorLogPath } = getLogFilePaths(
+      projectFolder,
+      logFilenameIdentifier,
+      logFolder
+    );
+    this.logPath = logPath;
+    this.errorLogPath = errorLogPath;
+    this.relativeLogPath = relativeLogPath;
+    this.relativeErrorLogPath = relativeErrorLogPath;
+
+    if (legacyLogPath !== this.logPath) {
+      FileSystem.deleteFile(this.logPath);
+    }
+
+    if (legacyErrorLogPath !== this.errorLogPath) {
+      FileSystem.deleteFile(this.errorLogPath);
+    }
 
     this._logWriter = FileWriter.open(this.logPath);
   }
